requirements.txt
fire==0.5.0
inject
langchain==0.0.148
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.2
